+++ src/Game.h
@@ -8,8 +8,8 @@
 #include "ECS/Design/Managers/ECSManager.h"
 #include <memory>
 
-const int FPS = 60;
-const int MILLISECS_PER_FRAME = 1000 / FPS;
+const size_t FPS = 60;
+const size_t MILLISECS_PER_FRAME = 1000 / FPS;
 
 class Game{
 
@@ -19,7 +19,7 @@ private:
         std::unique_ptr<Renderer> renderer;
         std::unique_ptr<InputProcessor> inputProcessor;
         ECSManager& manager;
-        int millisecsPreviousFrame = 0;
+        size_t millisecsPreviousFrame = 0;
 
 
 public:
